+++ packages/server/src/adapters/node-http/incomingMessageToRequest.ts
@@ -1,8 +1,5 @@
 import type * as http from 'http';
 import { TRPCError } from '../../@trpc/server';
-// eslint-disable-next-line no-restricted-imports
-import { isAbortError } from '../../unstable-core-do-not-import/http/isAbortError';
-import type { NodeHTTPResponse } from './types';
 
 export interface IncomingMessageWithBody extends http.IncomingMessage {
   /**
@@ -165,94 +162,4 @@ export function incomingMessageToRequest(
   const request = new Request(url, init);
 
   return request;
-}
-
-async function writeChunk(res: NodeHTTPResponse, chunk: Uint8Array) {
-  if (res.write(chunk) === false) {
-    await new Promise<void>((resolve, reject) => {
-      const onError = (err: unknown) => {
-        reject(err);
-        cleanup();
-      };
-      const onDrain = () => {
-        resolve();
-        cleanup();
-      };
-      const cleanup = () => {
-        res.off('error', onError);
-        res.off('drain', onDrain);
-      };
-      res.once('error', onError);
-      res.once('drain', onDrain);
-    });
-  }
-}
-
-/**
- * @internal
- */
-export async function writeBody(opts: {
-  res: NodeHTTPResponse;
-  signal: AbortSignal;
-  body: NonNullable<Response['body']>;
-}) {
-  const { res } = opts;
-
-  try {
-    const writableStream = new WritableStream({
-      async write(chunk) {
-        await writeChunk(res, chunk);
-        res.flush?.();
-      },
-      abort() {
-        if (!res.headersSent) {
-          res.statusCode = 500;
-        }
-      },
-    });
-
-    await opts.body.pipeTo(writableStream, {
-      signal: opts.signal,
-    });
-  } catch (err) {
-    if (isAbortError(err)) {
-      return;
-    }
-    throw err;
-  }
-}
-
-/**
- * @internal
- */
-export async function writeResponseToNodeHTTPResponse(opts: {
-  request: Request;
-  response: Response;
-  res: NodeHTTPResponse;
-}) {
-  const { response, res } = opts;
-
-  // Only override status code if it hasn't been explicitly set in a procedure etc
-  if (res.statusCode === 200) {
-    res.statusCode = response.status;
-  }
-  for (const [key, value] of response.headers) {
-    res.setHeader(key, value);
-  }
-  try {
-    if (response.body) {
-      await writeBody({
-        res,
-        signal: opts.request.signal,
-        body: response.body,
-      });
-    }
-  } catch (err) {
-    if (!res.headersSent) {
-      res.statusCode = 500;
-    }
-    throw err;
-  } finally {
-    res.end();
-  }
 }
