+++ packages/server/src/adapters/node-http/writeResponse.ts
@@ -1,0 +1,96 @@
+// eslint-disable-next-line no-restricted-imports
+import { isAbortError } from '../../unstable-core-do-not-import';
+import type { NodeHTTPResponse } from './types';
+
+async function writeResponseBodyChunk(
+  res: NodeHTTPResponse,
+  chunk: Uint8Array,
+) {
+  if (res.write(chunk) === false) {
+    await new Promise<void>((resolve, reject) => {
+      const onError = (err: unknown) => {
+        reject(err);
+        cleanup();
+      };
+      const onDrain = () => {
+        resolve();
+        cleanup();
+      };
+      const cleanup = () => {
+        res.off('error', onError);
+        res.off('drain', onDrain);
+      };
+      res.once('error', onError);
+      res.once('drain', onDrain);
+    });
+  }
+}
+/**
+ * @internal
+ */
+
+export async function writeResponseBody(opts: {
+  res: NodeHTTPResponse;
+  signal: AbortSignal;
+  body: NonNullable<Response['body']>;
+}) {
+  const { res } = opts;
+
+  try {
+    const writableStream = new WritableStream({
+      async write(chunk) {
+        await writeResponseBodyChunk(res, chunk);
+        res.flush?.();
+      },
+      abort() {
+        if (!res.headersSent) {
+          res.statusCode = 500;
+        }
+      },
+    });
+
+    await opts.body.pipeTo(writableStream, {
+      signal: opts.signal,
+    });
+  } catch (err) {
+    if (isAbortError(err)) {
+      return;
+    }
+    throw err;
+  }
+}
+/**
+ * @internal
+ */
+
+export async function writeResponse(opts: {
+  request: Request;
+  response: Response;
+  res: NodeHTTPResponse;
+}) {
+  const { response, res } = opts;
+
+  // Only override status code if it hasn't been explicitly set in a procedure etc
+  if (res.statusCode === 200) {
+    res.statusCode = response.status;
+  }
+  for (const [key, value] of response.headers) {
+    res.setHeader(key, value);
+  }
+  try {
+    if (response.body) {
+      await writeResponseBody({
+        res,
+        signal: opts.request.signal,
+        body: response.body,
+      });
+    }
+  } catch (err) {
+    if (!res.headersSent) {
+      res.statusCode = 500;
+    }
+    throw err;
+  } finally {
+    res.end();
+  }
+}
